requirements.txt
Pillow==2.4.0
PyVirtualDisplay==0.1.3
python-decouple==2.2
pyzmq==14.4.1
selenium==2.40.0
pyscreenshot==0.3.2
